+++ sdk/subscription/azure-mgmt-subscription/azure/mgmt/subscription/subscription_client.py
@@ -14,8 +14,6 @@ from msrest import Serializer, Deserializer
 from msrestazure import AzureConfiguration
 from .version import VERSION
 from .operations.operations import Operations
-from .operations.subscription_operation_operations import SubscriptionOperationOperations
-from .operations.subscription_factory_operations import SubscriptionFactoryOperations
 from .operations.subscriptions_operations import SubscriptionsOperations
 from .operations.tenants_operations import TenantsOperations
 from . import models
@@ -29,14 +27,18 @@ class SubscriptionClientConfiguration(AzureConfiguration):
     :param credentials: Credentials needed for the client to connect to Azure.
     :type credentials: :mod:`A msrestazure Credentials
      object<msrestazure.azure_active_directory>`
+    :param subscription_id: Subscription Id.
+    :type subscription_id: str
     :param str base_url: Service URL
     """
 
     def __init__(
-            self, credentials, base_url=None):
+            self, credentials, subscription_id, base_url=None):
 
         if credentials is None:
             raise ValueError("Parameter 'credentials' must not be None.")
+        if subscription_id is None:
+            raise ValueError("Parameter 'subscription_id' must not be None.")
         if not base_url:
             base_url = 'https://management.azure.com'
 
@@ -46,6 +48,7 @@ class SubscriptionClientConfiguration(AzureConfiguration):
         self.add_user_agent('Azure-SDK-For-Python')
 
         self.credentials = credentials
+        self.subscription_id = subscription_id
 
 
 class SubscriptionClient(SDKClient):
@@ -56,10 +59,6 @@ class SubscriptionClient(SDKClient):
 
     :ivar operations: Operations operations
     :vartype operations: azure.mgmt.subscription.operations.Operations
-    :ivar subscription_operation: SubscriptionOperation operations
-    :vartype subscription_operation: azure.mgmt.subscription.operations.SubscriptionOperationOperations
-    :ivar subscription_factory: SubscriptionFactory operations
-    :vartype subscription_factory: azure.mgmt.subscription.operations.SubscriptionFactoryOperations
     :ivar subscriptions: Subscriptions operations
     :vartype subscriptions: azure.mgmt.subscription.operations.SubscriptionsOperations
     :ivar tenants: Tenants operations
@@ -68,13 +67,15 @@ class SubscriptionClient(SDKClient):
     :param credentials: Credentials needed for the client to connect to Azure.
     :type credentials: :mod:`A msrestazure Credentials
      object<msrestazure.azure_active_directory>`
+    :param subscription_id: Subscription Id.
+    :type subscription_id: str
     :param str base_url: Service URL
     """
 
     def __init__(
-            self, credentials, base_url=None):
+            self, credentials, subscription_id, base_url=None):
 
-        self.config = SubscriptionClientConfiguration(credentials, base_url)
+        self.config = SubscriptionClientConfiguration(credentials, subscription_id, base_url)
         super(SubscriptionClient, self).__init__(self.config.credentials, self.config)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
@@ -82,10 +83,6 @@ class SubscriptionClient(SDKClient):
         self._deserialize = Deserializer(client_models)
 
         self.operations = Operations(
-            self._client, self.config, self._serialize, self._deserialize)
-        self.subscription_operation = SubscriptionOperationOperations(
-            self._client, self.config, self._serialize, self._deserialize)
-        self.subscription_factory = SubscriptionFactoryOperations(
             self._client, self.config, self._serialize, self._deserialize)
         self.subscriptions = SubscriptionsOperations(
             self._client, self.config, self._serialize, self._deserialize)